new_horizons/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==2.1.*
django-admin-honeypot==1.1.0
django-ckeditor==5.9.0
django-ckeditor-5==0.0.12
django-jet==1.0.8
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-rosetta==0.9.4
django-seed==0.2.2
Faker==4.1.2
idna==2.10
Pillow==7.2.0
polib==1.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
google-api-python-client==1.4.1
oauth2client==2.2.0